refactor(service): extract DNS logic into service_dns.go
