+++ daemon/protocol/helpers.go
@@ -42,7 +42,7 @@ func (p *Protocol) createSettingsResponse() *types.SettingsResp {
 		UserPrefs:                   prefs.UserPrefs,
 		WiFi:                        prefs.WiFiControl,
 		IsLogging:                   prefs.IsLogging,
-		AntiTracker:                 p._service.GetAntiTrackerStatus(),
+		AntiTracker:                 p._service.GetSettingsAntiTracker(),
 		// TODO: implement the rest of daemon settings
 	}
 }
@@ -99,7 +99,7 @@ func (p *Protocol) createConnectedResponse(state vpn.StateInfo) *types.Connected
 		ServerPort:      state.ServerPort,
 		VpnType:         state.VpnType,
 		ExitHostname:    state.ExitHostname,
-		Dns:             types.DnsStatus{Dns: manualDns, AntiTrackerStatus: p._service.GetAntiTrackerStatus()},
+		Dns:             types.DnsStatus{Dns: manualDns, AntiTrackerStatus: p._service.GetSettingsAntiTracker()},
 		IsTCP:           state.IsTCP,
 		Mtu:             state.Mtu,
 		V2RayProxy:      state.V2RayProxy,
