+++ daemon/service/service.go
@@ -795,244 +795,6 @@ func (s *Service) PausedTill() time.Time {
 	return s._pause._pauseTill
 }
 
-func (s *Service) saveDefaultDnsParams(dnsCfg dns.DnsSettings, antiTrackerCfg types.AntiTrackerMetadata) (retErr error) {
-	defaultParams := s.GetConnectionParams()
-
-	if defaultParams.ManualDNS.Equal(dnsCfg) && defaultParams.Metadata.AntiTracker.Equal(antiTrackerCfg) {
-		return nil
-	}
-
-	// save DNS and AntiTracker default metadata
-	defaultParams.ManualDNS = dnsCfg
-	defaultParams.Metadata.AntiTracker = antiTrackerCfg
-
-	return s.setConnectionParams(defaultParams)
-}
-
-// GetActiveDNS() returns DNS active settings for current VPN connection:
-// - if 'antiTracker' is enabled - returns DNS of AntiTracker server
-// - else if manual DNS is defined - returns manual DNS
-// - else returns default DNS configuration for current VPN connection
-// *Note! If VPN disconnected - returns empty data
-func (s *Service) GetActiveDNS() (dnsCfg dns.DnsSettings, err error) {
-	vpnObj := s._vpn
-	if vpnObj == nil {
-		return dns.DnsSettings{}, nil //VPN DISCONNECTED
-	}
-
-	_, _, manualDns, err := s.GetDefaultManualDnsParams()
-	if err != nil {
-		return dns.DnsSettings{}, err
-	}
-	if !manualDns.IsEmpty() {
-		return manualDns, nil
-	}
-
-	return dns.DnsSettingsCreate(vpnObj.DefaultDNS()), nil
-}
-
-// GetDefaultManualDnsParams returns default manual DNS parameters
-// Returns:
-//
-//	manualDnsCfg - default manual DNS parameters
-//	antiTrackerCfg - default AntiTracker parameters
-//	realDnsValue - real DNS value (if 'antiTracker' is enabled - it will contain DNS of AntiTracker server)
-func (s *Service) GetDefaultManualDnsParams() (manualDnsCfg dns.DnsSettings, antiTrackerCfg types.AntiTrackerMetadata, realDnsValue dns.DnsSettings, err error) {
-	defaultParams := s.GetConnectionParams()
-
-	manualDnsCfg = defaultParams.ManualDNS
-	realDnsValue = defaultParams.ManualDNS
-	antiTrackerCfg = defaultParams.Metadata.AntiTracker
-
-	if antiTrackerCfg.Enabled {
-		realDnsValue, err = s.getAntiTrackerDns(antiTrackerCfg.Hardcore, antiTrackerCfg.AntiTrackerBlockListName)
-	}
-
-	return manualDnsCfg, antiTrackerCfg, realDnsValue, err
-}
-
-// SetManualDNS update default DNS parameters AND apply new DNS value for current VPN connection
-// If 'antiTracker' is enabled - the 'dnsCfg' will be ignored
-func (s *Service) SetManualDNS(dnsCfg dns.DnsSettings, antiTracker types.AntiTrackerMetadata) (changedDns dns.DnsSettings, retErr error) {
-	prefs := s.Preferences()
-	if !dnsCfg.IsEmpty() || antiTracker.Enabled {
-		if prefs.IsInverseSplitTunneling() && prefs.SplitTunnelAnyDns {
-			return dns.DnsSettings{}, fmt.Errorf("custom DNS or AntiTracker cannot be enabled while allowing all DNS for Inverse Split Tunnel mode; please block non-IVPN DNS first in the Inverse Split Tunnel configuration")
-		}
-	}
-
-	isChanged := false
-	defer func() {
-		if isChanged {
-			// Apply Firewall rule (for Inverse Split Tunnel): allow DNS requests only to IVPN servrers or to manually defined server
-			if err := s.splitTunnelling_ApplyConfig(); err != nil {
-				log.Error(err)
-			}
-		}
-	}()
-
-	// Update default metadata
-	defaultParams := s.GetConnectionParams()
-	// save DNS and AntiTracker default metadata
-	if !defaultParams.ManualDNS.Equal(dnsCfg) {
-		defaultParams.ManualDNS = dnsCfg
-		isChanged = true
-	}
-	if !defaultParams.Metadata.AntiTracker.Equal(antiTracker) {
-		at, err := s.normalizeAntiTrackerBlockListName(antiTracker)
-		if err != nil {
-			return changedDns, err
-		}
-		defaultParams.Metadata.AntiTracker = at
-		isChanged = true
-	}
-	if isChanged {
-		s.setConnectionParams(defaultParams)
-	}
-
-	// Get anti-tracker DNS settings
-	changedDns = dnsCfg
-	if antiTracker.Enabled {
-		atDns, err := s.getAntiTrackerDns(antiTracker.Hardcore, antiTracker.AntiTrackerBlockListName)
-		if err != err {
-			return dns.DnsSettings{}, err
-		}
-		changedDns = atDns
-	}
-
-	vpn := s._vpn
-	if vpn == nil {
-		// no active VPN connection
-		return changedDns, nil
-	}
-
-	if dnsCfg.IsEmpty() && !antiTracker.Enabled {
-		return dns.DnsSettings{}, vpn.ResetManualDNS()
-	}
-	return changedDns, vpn.SetManualDNS(changedDns)
-}
-
-func (s *Service) GetManualDNSStatus() dns.DnsSettings {
-	return s.GetConnectionParams().ManualDNS
-}
-
-func (s *Service) GetAntiTrackerStatus() types.AntiTrackerMetadata {
-	// Get AntiTracker DNS settings. If error - use default date and ignore error
-	retAtMetadata, err := s.normalizeAntiTrackerBlockListName(s.GetConnectionParams().Metadata.AntiTracker)
-	if err != nil {
-		log.Error(fmt.Sprintf("failed to normalize AntiTracker block list name: %v (using '%s')", err, retAtMetadata.AntiTrackerBlockListName))
-	}
-	return retAtMetadata
-}
-
-// Normze AntiTracker block list name:
-// - if antiTrackerPlusList not defined - return default value
-// - if antiTrackerPlusList defined - check if it is valid; if not valid - return default value and error
-func (s *Service) normalizeAntiTrackerBlockListName(antiTracker types.AntiTrackerMetadata) (types.AntiTrackerMetadata, error) {
-	var retError error
-
-	atBlistName := strings.ToLower(strings.TrimSpace(antiTracker.AntiTrackerBlockListName))
-	// check if block list name is known
-	if atBlistName != "" {
-		servers, err := s.ServersList()
-		if err == nil {
-			for _, atp_svr := range servers.Config.AntiTrackerPlus.DnsServers {
-				if strings.ToLower(strings.TrimSpace(atp_svr.Name)) == atBlistName {
-					// Block-list name is OK. Just ensure to use correct case
-					antiTracker.AntiTrackerBlockListName = strings.TrimSpace(atp_svr.Name)
-					return antiTracker, nil
-				}
-			}
-		}
-
-		retError = fmt.Errorf("unexpected DNS block list name: '%s'", antiTracker.AntiTrackerBlockListName)
-	}
-
-	// Set default block list name (if empty)
-	if tmpDns, err := s.getAntiTrackerDns(antiTracker.Hardcore, ""); err == nil {
-		if tmpAt, err := s.getAntiTrackerInfo(tmpDns); err == nil {
-			antiTracker.AntiTrackerBlockListName = tmpAt.AntiTrackerBlockListName
-		}
-	}
-
-	return antiTracker, retError
-}
-
-// Get DNS server according to AntiTracker parameters
-func (s *Service) getAntiTrackerDns(isHardcore bool, antiTrackerPlusList string) (dnsCfg dns.DnsSettings, err error) {
-	defer func() {
-		if dnsCfg.IsEmpty() && err == nil {
-			err = fmt.Errorf("unable to determine AntiTracker DNS")
-		}
-	}()
-	servers, err := s.ServersList()
-	if err != nil {
-		return dns.DnsSettings{}, fmt.Errorf("failed to determine AntiTracker parameters: %w", err)
-	}
-
-	// AntiTracker Plus list
-	atListName := strings.ToLower(strings.TrimSpace(antiTrackerPlusList))
-	if len(atListName) == 0 {
-		// if block list name not defined - use default AntiTracker block list "Basic"
-		atListName = "basic"
-	}
-
-	if len(atListName) > 0 {
-		for _, atp_svr := range servers.Config.AntiTrackerPlus.DnsServers {
-			if strings.ToLower(strings.TrimSpace(atp_svr.Name)) == atListName {
-				if isHardcore {
-					return dns.DnsSettingsCreate(net.ParseIP(atp_svr.Hardcore)), nil
-				}
-				return dns.DnsSettingsCreate(net.ParseIP(atp_svr.Normal)), nil
-			}
-		}
-	}
-
-	// If AntiTracker Plus block list not found - ignore 'antiTrackerPlusList' and use old-style AntiTracker DNS
-	if isHardcore {
-		return dns.DnsSettingsCreate(net.ParseIP(servers.Config.Antitracker.Hardcore.IP)), nil
-	}
-	return dns.DnsSettingsCreate(net.ParseIP(servers.Config.Antitracker.Default.IP)), nil
-}
-
-// Get AntiTracker info according to DNS settings
-func (s *Service) getAntiTrackerInfo(dnsVal dns.DnsSettings) (types.AntiTrackerMetadata, error) {
-	// If uses encrypted DNS or multiple DNS servers - it is custom DNS, no AntiTracker info
-	if dnsVal.IsEmpty() || dnsVal.UseEncryption() || len(dnsVal.Servers) != 1 {
-		return types.AntiTrackerMetadata{}, nil
-	}
-
-	servers, err := s.ServersList()
-	if err != nil {
-		return types.AntiTrackerMetadata{}, fmt.Errorf("failed to determine AntiTracker parameters: %w", err)
-	}
-
-	dnsHost := strings.ToLower(strings.TrimSpace(dnsVal.Servers[0].Address))
-	if dnsHost == "" {
-		return types.AntiTrackerMetadata{}, nil
-	}
-
-	// Check AntiTracker Plus lists
-	for _, atp_svr := range servers.Config.AntiTrackerPlus.DnsServers {
-		if strings.EqualFold(dnsHost, strings.TrimSpace(atp_svr.Normal)) {
-			return types.AntiTrackerMetadata{Enabled: true, Hardcore: false, AntiTrackerBlockListName: atp_svr.Name}, nil
-		}
-		if strings.EqualFold(dnsHost, strings.TrimSpace(atp_svr.Hardcore)) {
-			return types.AntiTrackerMetadata{Enabled: true, Hardcore: true, AntiTrackerBlockListName: atp_svr.Name}, nil
-		}
-	}
-
-	// Check AntiTracker values
-	if strings.EqualFold(dnsHost, strings.TrimSpace(servers.Config.Antitracker.Default.IP)) {
-		return types.AntiTrackerMetadata{Enabled: true, Hardcore: false}, nil
-	}
-	if strings.EqualFold(dnsHost, strings.TrimSpace(servers.Config.Antitracker.Hardcore.IP)) {
-		return types.AntiTrackerMetadata{Enabled: true, Hardcore: true}, nil
-	}
-
-	return types.AntiTrackerMetadata{}, nil
-}
-
 // ////////////////////////////////////////////////////////
 // KillSwitch
 // ////////////////////////////////////////////////////////
@@ -1476,7 +1238,7 @@ func (s *Service) splitTunnelling_ApplyConfig() (retError error) {
 	}
 	isVpnConnected := s.Connected() && !s.IsPaused()
 	if isVpnConnected && prefs.IsInverseSplitTunneling() && !prefs.SplitTunnelAnyDns {
-		dnsCfg, err := s.GetActiveDNS() // returns nil when VPN not connected
+		dnsCfg, err := s.getActiveDNS() // returns empty DNS when VPN not connected
 		if err != nil {
 			return fmt.Errorf("failed to apply the firewall rule to allow DNS requests only to the IVPN server: %w", err)
 		}
